refactor: split DNSMessage into a separate target
