+++ Tests/DNSClientTests/DNSUDPClientTests.swift
@@ -1,4 +1,5 @@
 import DNSClient
+import DNSMessage
 import NIO
 import XCTest
